+++ examples/benchmark/src/alltoallv/alltoallv_strategy.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 struct alltoallv_strategy_impl {
-    size_t comm_size = 0;
+    int comm_size = 0;
     std::vector<size_t> send_counts;
     std::vector<size_t> recv_counts;
 
@@ -31,7 +31,7 @@ struct alltoallv_strategy_impl {
                         const bench_exec_attr& bench_attr,
                         req_list_t& reqs,
                         Args&&... args) {
-        for (size_t idx = 0; idx < comm_size; idx++) {
+        for (int idx = 0; idx < comm_size; idx++) {
             send_counts[idx] = count;
             recv_counts[idx] = count;
         }
