+++ examples/benchmark/src/reduce_scatter/sycl_reduce_scatter_coll.hpp
@@ -24,7 +24,7 @@ struct sycl_reduce_scatter_coll : sycl_base_coll<Dtype, reduce_scatter_strategy_
         if (base_coll::get_sycl_mem_type() != SYCL_MEM_BUF)
             return;
 
-        size_t local_rank = comm.rank();
+        int local_rank = comm.rank();
         for (size_t b_idx = 0; b_idx < base_coll::get_buf_count(); b_idx++) {
             stream.get_native().submit([&](handler& h) {
                 auto send_buf = (static_cast<sycl_buffer_t<Dtype>*>(send_bufs[b_idx][rank_idx]));
