+++ examples/cpu/cpu_allreduce_bf16_test.cpp
@@ -13,7 +13,7 @@
 #define CHECK_ERROR(send_buf, recv_buf, comm) \
     { \
         /* https://www.mcs.anl.gov/papers/P4093-0713_1.pdf */ \
-        size_t comm_size = comm.size(); \
+        int comm_size = comm.size(); \
         double log_base2 = log(comm_size) / log(2); \
         double g = (log_base2 * BF16_PRECISION) / (1 - (log_base2 * BF16_PRECISION)); \
         for (size_t i = 0; i < COUNT; i++) { \
@@ -21,7 +21,7 @@
             double max_error = g * expected; \
             if (fabs(max_error) < fabs(expected - recv_buf[i])) { \
                 printf( \
-                    "[%zu] got recv_buf[%zu] = %0.7f, but expected = %0.7f, max_error = %0.16f\n", \
+                    "[%d] got recv_buf[%zu] = %0.7f, but expected = %0.7f, max_error = %0.16f\n", \
                     comm.rank(), \
                     i, \
                     recv_buf[i], \
