+++ internal/examples/l0/alltoallv_l0.cpp
@@ -30,7 +30,7 @@ template <class processing_type>
 void user_thread_idx(size_t thread_idx,
                      const std::vector<std::pair<size_t, cl::sycl::device>>& devices,
                      cl::sycl::context ctx,
-                     size_t total_devices_in_cluster,
+                     int total_devices_in_cluster,
                      std::shared_ptr<ccl::kvs_interface> kvs_instance) {
     using namespace ::native;
 
@@ -75,7 +75,7 @@ void user_thread_idx(size_t thread_idx,
     for (auto& comm : comms) {
         // get native l0* /
         ccl::communicator::device_type dev = comm.get_device().get_native();
-        size_t rank = comm.rank();
+        int rank = comm.rank();
 
         // create comm split attr
         auto device_spilt_attr = ccl::create_comm_split_attr();
@@ -117,7 +117,7 @@ void user_thread_idx(size_t thread_idx,
     //allreduce
     std::vector<ccl::event> reqs;
     for (auto& comm : comms) {
-        size_t rank = comm.rank();
+        int rank = comm.rank();
 
         /*
         if (!comm.is_ready())
@@ -151,7 +151,7 @@ void user_thread_idx(size_t thread_idx,
     {
         std::unique_lock<std::mutex> lock(printout_mutex);
         for (auto& dev_it : memory_storage) {
-            size_t rank = dev_it.first;
+            int rank = dev_it.first;
             const auto& handles = dev_it.second;
             std::cout << "rank : " << rank << std::endl;
             for (const auto& mem : handles) {
@@ -168,7 +168,7 @@ template <class processing_type>
 void user_thread_idx(size_t thread_idx,
                      std::vector<std::pair<size_t, ccl::device_index_type>> ranked_device_indices,
                      std::shared_ptr<::native::ccl_context> ctx,
-                     size_t total_devices_in_cluster,
+                     int total_devices_in_cluster,
                      std::shared_ptr<ccl::kvs_interface> kvs) {
     using namespace ::native;
 
@@ -228,7 +228,7 @@ void user_thread_idx(size_t thread_idx,
     for (auto& comm : comms) {
         // get native l0*
         ccl::communicator::device_native_reference_t dev = comm->get_device();
-        size_t rank = comm->rank();
+        int rank = comm->rank();
 
         // wrapped L0-native API for devices: create native buffers
         auto mem_send = dev->alloc_memory<processing_type>(COUNT * local_topology_size,
@@ -265,7 +265,7 @@ void user_thread_idx(size_t thread_idx,
     std::vector<std::shared_ptr<ccl::event>> reqs;
     ccl::attr attr{};
     for (auto& comm : comms) {
-        size_t rank = comm->rank();
+        int rank = comm->rank();
 
         if (!comm->is_ready()) {
             std::cerr << "Communicator by rank: " << rank << " should be ready already"
@@ -294,7 +294,7 @@ void user_thread_idx(size_t thread_idx,
     {
         std::unique_lock<std::mutex> lock(printout_mutex);
         for (auto& dev_it : memory_storage) {
-            size_t rank = dev_it.first;
+            int rank = dev_it.first;
             const auto& handles = dev_it.second;
             std::cout << "rank : " << rank << std::endl;
             for (const auto& mem : handles) {
